Fix warning due to unsigned / signed comparison
git-svn-id: http://root.cern.ch/svn/root/trunk@22584 27541ba8-7e3a-0410-8455-c3a389f83636

+++ proofd/src/XrdProofSched.cxx
@@ -210,7 +210,7 @@ int XrdProofSched::GetNumWorkers(XrdProofServProxy *xps)
       TRACE(DBG, "GetNumWorkers: "<< (*iter)->fImage<<
                  " : # act: "<<(*iter)->fProofServs.size());
       if ((*iter)->fType != 'M'
-         && (*iter)->fProofServs.size() < fOptWrksPerUnit)
+         && (int)((*iter)->fProofServs.size()) < fOptWrksPerUnit)
          nFreeCPUs++;
    }
